Sort memcache_hosts return value.
The list of memcache_hosts is typically rendered in a config file
template, and if restarting the service hinges on config file changes,
the fact that memcache_hosts could return the same set of values in a
different order can result in unwanted restarts.

Simply sort the return value to ensure that given the same set of
'private-address' values, the result's order will not change.